Possibility to include terrain elevation in ground track termination
Added terrain to bool aboveSurface. oapiSurfaceElevation is very slow, so is off by default.
Also add to show both analytic and numeric ground track in debug mode.

Note that there is a bug with the track when passing perigee. Probably from the stepTrA again :'(
Considering to remove it all, and instead add a simple constant timestep.